added sprite flags xflip yflip pallete

+++ Source/Common.hpp
@@ -10,12 +10,12 @@ using owner = T;
 
 constexpr size_t operator""_Kib(unsigned long long kibs)
 {
-	return kibs * 1024;
+	return static_cast<size_t>(kibs * 1024);
 }
 
 constexpr size_t operator""_Mib(unsigned long long mibs)
 {
-	return mibs * 1024 * 1024;
+	return static_cast<size_t>(mibs * 1024 * 1024);
 }
 
 template<class F>
